+++ advanced_template/special_scripts/__init__.py
@@ -1,0 +1,7 @@
+"""
+This allows me to import from the special_scripts folder instead of importing from the selected file.
+Used to be: from advanced_template.special_scripts.[Filename] import [Classname]
+Now is    : from advanced_template.views import [Classname]
+"""
+
+from .buttons import BasicButton